+++ mobile/android/android-components/components/feature/awesomebar/src/main/java/mozilla/components/feature/awesomebar/AwesomeBarFeature.kt
@@ -252,6 +252,6 @@ internal class ToolbarEditListener(
     }
 
     override fun onCancelEditing(): Boolean {
-        return true
+        return !inputStarted
     }
 }